C++ code for Best fit,worst fit and first fit algorithm for memory management in OS
