+++ Memory_management_algorithm/worst_fit.cpp
@@ -1,6 +1,8 @@
+#include <bits/stdc++.h>
 #include <stdio.h>
 #include <stdlib.h>
 
+using namespace std;
 
 void first_fit(int memory_block[],int process_size[],int m,int n)
 {
@@ -77,14 +79,34 @@ void first_fit(int memory_block[],int process_size[],int m,int n)
 
 int main()
 {
+	int num_of_mem_blocks;
+	int num_of_processes;
 
-	int memory_block[] = {100,500,250,300,50,400};
+	cout<<"Enter the number of memory blocks\n";
 
-	int process_size[] = {300,250,229,97,103,350};
+	cin>>num_of_mem_blocks;
 
-	int num_of_mem_blocks = sizeof(memory_block)/sizeof(memory_block[0]);
+	cout<<"Enter the number of processes\n";
 
-	int num_of_processes = sizeof(process_size)/sizeof(process_size[0]);
+	cin>>num_of_processes;
+
+	int memory_block[num_of_mem_blocks];
+
+	int process_size[num_of_processes];
+
+	cout<<"Enter the memory blocks size:\n";
+
+	for(int i=0;i<num_of_mem_blocks;i++)
+	{
+		cin>>memory_block[i];
+	}
+
+	cout<<"Enter the process size:\n";
+
+	for(int i=0;i<num_of_processes;i++)
+	{
+		cin>>process_size[i];
+	}
 
 	first_fit(memory_block,process_size,num_of_mem_blocks,num_of_processes);
 
@@ -92,4 +114,28 @@ int main()
 }
 
 
-
+/*    INPUT:   
+ *
+ *    		Memory Blocks:
+ *
+ *    		100 500 200 300 600
+ *
+ *    		Processes:
+ *
+ *    		212 417 112 426
+ *
+ *
+ *    	OUTPUT:
+ *
+ *    	Proesses              Memory block
+ *
+ *    	212			600
+ *
+ *    	417			500
+ *
+ *    	112			300
+ *
+ *    	426			Not Allocated
+ *
+ *
+ */
